app/requirements.txt
fastapi
pydantic
datetime
mysql.connector
requests